Update C64.emu to VICE 3.4 with additional tweaks
* Imagine: Full path in Base::libPath() isn't needed on Android >= 7.0
* EmuFramework: Add EmuControls::updateVControllerMapping()
* EmuFramework: Add EmuApp::launchSystem()
* EmuFramework: Add SystemOptionView::pushAndShowFirmwareFilePathMenu() with support for using an archive file as the firmware path
* C64.emu: Fix memory leak in vdrive_device_setup()
* C64.emu: Add prompt to overwrite existing disk image when starting with a blank disk
* C64.emu: Check for plugin's main ROM file before calling init_main()
* C64.emu: Only attempt autostart if plugin has autostart_autodetect()
* C64.emu: Check that all plugin's symbols are defined at load time
* C64.emu: Remove hard-coded keyboard codes and use system-specific gtk3_sym.vkm file as an intermediate mapping
* C64.emu: Try asset path for system files as a fallback
* C64.emu: Define all external functions used by plugins in their own headers
* C64.emu: Don't enumerate plugins based on file path

+++ C64.emu/src/vice/alarm.h
@@ -160,7 +160,7 @@ inline static void alarm_set(alarm_t *alarm, CLOCK cpu_clk)
         /* Not pending yet: add.  */
 
         new_idx = (int)(context->num_pending_alarms);
-        if (unlikely(new_idx >= ALARM_CONTEXT_MAX_PENDING_ALARMS)) {
+        if (unlikely(new_idx >= (int)ALARM_CONTEXT_MAX_PENDING_ALARMS)) {
             alarm_log_too_many_alarms();
             return;
         }
