+++ C64.emu/src/vice/c64/patchrom.c
@@ -460,23 +460,25 @@ static const unsigned short patch_bytes[] = {
 
 int patch_rom_idx(int rev)
 {
-    WORD sum;                   /* ROM checksum */
+    uint16_t sum;                   /* ROM checksum */
     int curr, num;
     short bytes, n, i = 0;
-    WORD a;
+    uint16_t a;
 
     if (c64rom_get_kernal_chksum_id(&sum, &curr) < 0) {
-        log_error(LOG_DEFAULT, "ROM not patched: Unknown Kernal image.  ID: %d ($%02X) Sum: %d ($%04X).", curr, curr, sum, sum);
+        log_error(LOG_DEFAULT,
+                "ROM not patched: Unknown Kernal image.  ID: %d ($%02X) Sum: %d ($%04X).",
+                curr, (unsigned int)curr, sum, sum);
         return -1;
     }
-    log_verbose("Trying Kernal ROM patch: id:%d/sum:%d to id:%d.", curr, sum, rev);
+    log_verbose("Trying Kernal ROM patch: id:%d/sum:%u to id:%d.", curr, sum, rev);
 
     if (rev == curr) {
         log_verbose("ROM not patched: Already revision #%d.", curr);
         return 0;
     }
     if (rev < 0) {
-        log_error(LOG_DEFAULT, "ROM not patched: Unknown ID: %d ($%02X).", rev, rev);
+        log_error(LOG_DEFAULT, "ROM not patched: Unknown ID: %d ($%02X).", rev, (unsigned int)rev);
         return -1;
     }
 
@@ -509,13 +511,13 @@ int patch_rom_idx(int rev)
 
     i = 0;
     while ((bytes = patch_bytes[i++]) > 0) {
-        a = (WORD)patch_bytes[i++];
+        a = (uint16_t)patch_bytes[i++];
 
         log_verbose("  %.4X (%d byte%s)", a & 0xFFFF, bytes, ((bytes > 1) ? "s" : ""));
 
         i += (bytes * rev);     /* select patch */
         for (n = bytes; n--; ) {
-            c64memrom_rom64_store(a++, (BYTE)patch_bytes[i++]);
+            c64memrom_rom64_store(a++, (uint8_t)patch_bytes[i++]);
         }
 
         i += (bytes * (PATCH_VERSIONS - rev));  /* skip patch */
