+++ C64.emu/src/vice/resid/sid.cc
@@ -949,7 +949,7 @@ int SID::clock_resample(cycle_count& delta_t, short* buf, int n, int interleave)
     // Linear interpolation.
     // fir_offset_rmd is equal for all samples, it can thus be factorized out:
     // sum(v1 + rmd*(v2 - v1)) = sum(v1) + rmd*(sum(v2) - sum(v1))
-    int v = v1 + (fir_offset_rmd*(v2 - v1) >> FIXP_SHIFT);
+    int v = v1 + int((unsigned(fir_offset_rmd)*unsigned(v2 - v1)) >> FIXP_SHIFT);
 
     v >>= FIR_SHIFT;
